Task_2_Image_Classification/requirements.txt
numpy==1.24.4 

pandas==1.5.3 

torch==1.13.1 

torchvision==0.14.1 

transformers==4.35.0 

datasets==2.12.0 

peft==0.6.0 

accelerate==0.25.0 

huggingface_hub==0.17.3 

Pillow==9.5.0 

matplotlib==3.7.1 

jupyter==1.0.0 

seaborn==0.12.2


